# docs/swagger.yaml
definitions:
...
  models.TaskRequest:
    properties:
      actual_hours:
        type: number
      assignee_id:
        type: string
      category:
        description: BARU
        type: string
      completed:
        description: BARU
        type: boolean
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      priority:
        description: '''low'', ''medium'', ''high'''
        type: string
      progress:
        type: integer
      project_id:
        type: string
      status:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  models.ToggleTaskRequest:
    properties:
      completed:
        type: boolean
    type: object
  models.UpdateTaskCategoryRequest:
    properties:
      category:
        type: string
    required:
    - category
    type: object
...
  utils.GinResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: Get list of tasks assigned to the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Get all tasks for current user
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task for the current user
      parameters:
      - description: Task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Create new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get a specific task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update an existing task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{id}/category:
    patch:
      consumes:
      - application/json
      description: Update category of a specific task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Update task category
      tags:
      - tasks
  /tasks/{id}/complete:
    patch:
      consumes:
      - application/json
      description: Mark a task as completed
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Complete task
      tags:
      - tasks
  /tasks/{id}/progress:
    patch:
      consumes:
      - application/json
      description: Update progress of a specific task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Progress data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Update task progress
      tags:
      - tasks
  /tasks/{id}/toggle:
    patch:
      consumes:
      - application/json
      description: Toggle the completed status of a task (true/false)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Completion status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ToggleTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Toggle task completion status
      tags:
      - tasks
  /tasks/statistics:
    get:
      consumes:
      - application/json
      description: Get task statistics for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Get task statistics
      tags:
      - tasks
  /tasks/upcoming:
    get:
      consumes:
      - application/json
      description: Get upcoming tasks for dashboard (limit 3)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GinResponse'
      security:
      - BearerAuth: []
      summary: Get upcoming tasks
      tags:
      - tasks
...
swagger: "2.0"